# feegow-challenge-covid-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Vacinação de Funcionários",
        "description": "Uma API para registro de vacinação de funcionários contra a COVID-19.",
        "contact": {
            "name": "Suporte Técnico",
            "email": "suporte@empresa.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/funcionario-vacina/{funcionario_cpf}/{vacina_id}/{dose}": {
            "put": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Atualiza os detalhes de uma vacinação específica de um funcionário",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "funcionario_cpf",
                        "in": "path",
                        "description": "CPF do funcionário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "vacina_id",
                        "in": "path",
                        "description": "ID da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dose",
                        "in": "path",
                        "description": "Dose da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data_dose"
                                ],
                                "properties": {
                                    "data_dose": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-14"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro de vacinação atualizado com sucesso!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Remove uma vacinação específica de um funcionário",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "funcionario_cpf",
                        "in": "path",
                        "description": "CPF do funcionário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "vacina_id",
                        "in": "path",
                        "description": "ID da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dose",
                        "in": "path",
                        "description": "Dose da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Registro de vacinação removido com sucesso!"
                    }
                }
            }
        },
        "/api/vacinas": {
            "get": {
                "tags": [
                    "Vacinas"
                ],
                "summary": "Lista todas as vacinas",
                "operationId": "d988745b062519447330347bbd70d88b",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vacinas"
                ],
                "summary": "Cria uma nova vacina",
                "operationId": "cb971799af2d809c01e705e73d5c5a3b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "lote",
                                    "data_validade"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Vacina X"
                                    },
                                    "lote": {
                                        "type": "string",
                                        "example": "Lote123"
                                    },
                                    "data_validade": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vacina criada com sucesso"
                    }
                }
            }
        },
        "/api/vacinas/{id}": {
            "get": {
                "tags": [
                    "Vacinas"
                ],
                "summary": "Exibe uma vacina específica",
                "operationId": "64c90b8730bc822607c18d5ce3039d95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "404": {
                        "description": "Vacina não encontrada."
                    }
                }
            },
            "put": {
                "tags": [
                    "Vacinas"
                ],
                "summary": "Atualiza uma vacina específica",
                "operationId": "3695cd66443bb8dc1f30cd49cf8fc559",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "lote",
                                    "data_validade"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Nova Vacina X"
                                    },
                                    "lote": {
                                        "type": "string",
                                        "example": "NovoLote123"
                                    },
                                    "data_validade": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacina atualizada com sucesso"
                    },
                    "404": {
                        "description": "Vacina não encontrada."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vacinas"
                ],
                "summary": "Deleta uma vacina específica",
                "operationId": "4a0977481a7e367bfac3284855c70917",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vacina deletada com sucesso"
                    },
                    "404": {
                        "description": "Vacina não encontrada."
                    }
                }
            }
        }
    }
}